🔨 Add null handler to 'event'
The alloy::io::event class type-erases the underlying event type
using a fixed-buffer and a handler. In the case that the event is
constructed without an underlying event value, the handler was
previously null -- which necessitated checks everywhere that the
handler was used.

This nullptr discrepancy pessimizes the non-null use-case by forcing
these checks, even in a system where the events will always be
non-null. This was used so that things like the default "id" or
"priority" could be set to "0" and "none" respectively.

In order to avoid this pessimization, the null-case has been
internalized as a special "null_handler". This now prevents the
pessimization on the common path where values are set -- trading this
for indirection for null cases. This is a fair trade-off since
it is less likely for events to be "null" then they are to be
non-null.