🩹 Fix bad event constructor
The `event(Event&&)` constructor incorrectly uses `e.m_storage` as the
destination of placement-`new` invocations. This is incorrect, since
`e` is the input `Event` object -- which is not an `alloy::io::event`.

This has been corrected to just `m_storage`, which is the correct
destination for the placement-new invocation.