🩹 Make ALLOY_<COMP>_STATIC a public flag
The export/import macros used to produce usable symbols on Windows
was broken for static builds, due to the `ALLOY_<COMP>_STATIC`
flag only being defined during compilation, but not during consumption.

This was resulting in unit tests that used static libraries to build
while looking for an `__declspec(export)` symbol, without ever having
been built with a `__declspec(import)` definition.

To fix this issue, the `ALLOY_<COMP>_STATIC` symbols have been
propagated to `PUBLIC` visibility in CMake, so that both consumers and
builders may receive the same flag. This fixes the failure to locate
the symbol when building on Windows, and should also address builds
that make use of `-fvisibility=hidden` on Linux/macOS