✨ Add compiler intrinsics for different branches
WIth the newly added `ALLOY_HOT` and `ALLOY_COLD` macro, new utilities
have been added to the `alloy::core::compiler` intrinsic class for
adding paths that are assumed to be hot and paths that are assumed to
be cold.

Additionally new functions have been added for handling code that only
exist in debug and release mode. This code will still be checked for
syntactical accuracy, but will be optimized away when compiling with
`NDEBUG` for release, or without `NDEBUG` for debug mode.