🔨 Refactor assert handlers
The `ALLOY_ASSERT` macro has been updated to now support a
user-configurable assertion handler. To do this, the previously
inlined assertion handler has now been moved to its own translation
unit -- moving the breakpoint to be in the body of the assertion
lambda. The actual handler itself is documented to be a cold-path
to minimize impact on the caller so the icache doesn't get polluted.