🔨 Restructure intrinsics.hpp
The intrinsics structure used to be broken into a subdirectory where
each individual compiler received its own unique header. This was
asymmetric with the rest of the system, since these headers could only
be included by the appropriate compiler, but had the visibility of any
other header in the system.

To clean this up, the entirety of those individual headers have been
collapsed into the intrinsics.hpp header, and individual compiler
logic checks that was not used has been removed.