✨ And UTF-8 string literals
C++20 is introducing support for a new character type, char8_t, which is
guaranteed to be formatted in utf8. In addition to this, new semantic
types for string and string_view are added, the u8string and
u8string_view.

These have been introduced to alloy, and will conditionally expand to
char in C++17, or char8_t in C++20 and above. These leverage the
type-system changes, in that expressions of `u8'c'` will expand to a
`char` type in C++17 or a `char8_t` in C++20 -- which allow a simple,
portable way to test for the existence of C++20 and char8_t support
without explicitly feature-test macros or `__cplusplus` inspection --
which often breaks with MSVC.