🔨 Refactor vector headers
The various vector headers have been refactored so that definitions
that are large and not `constexpr` are now defined in the corresponding
`cpp` file instead.

The original design of having the entire `vector` being inline was
entirely due to `vector` originally being a template in previous
versions -- which required it to be inline. Since `vector` has been
modified to be fixed in recent versions, the larger functions are
actually detrimental for inlining since they pollute the icache and
may be infrequently used.

These functions have been relocated to the `cpp` file accordingly,
leaving only smaller `constexpr` functions to be define `inline`.