🩹 Fix visibility of vector classes
The various vector classes were previously refactored to have their
definitions spread throughout multiple `cpp` files. This has resulted
in a number of the symbols not being resolved when building on Windows.

To account for this, the `ALLOY_CORE_API` macro for exporting the
symbols has been added to each of the defined classes.