✨ Add /W4 and /WX to MSVC compile commands
The MSVC compile commands have been updated to now support the highest
warning level (`/W4`) in addition to erroring on warnings (`/WX`).

In order to do this, a policy has also been set to ensure that CMake
will not, by default, force `/W3` to be set -- which triggers warnings
by the compiler that both are defined.