🔨 Address MSVC's incorrect "shadow" warning
MSVC incorrectly identifies an if/else if/else statement using
automatic variables in its construction to be "shadowing" the
previous scope... Which it shouldn't be, since the scope is limited to
the interior of the branch -- and should not seep into the next
condition.

The code here has been refactored to make use of the switch/case
syntax, since this is now supported. Hopefully this fixes the bug.