🔨 Change event_id to integral value
The `event_id` type was previously defined to be a unique integral
value by making it an `enum class` with no discriminated values.

However, this was triggering warnings on clang with `-Wswitch` and
on MSVC as well. As a result, the event id types have been changed
to be numeric types instead to ensure that this will not error during
compilation.