✨ Add attribute_buffer_view utility
When dealing directly with underlying graphics APIs, there are often
aliasing issues to consider. APIs like OpenGL require a contiguous
sequence to floating-point values, but often in libraries we are working
in higher-level code that is actually dealing with Vectors, or Color
types. Passing sequences of these to an API call like OpenGL *may*
work (and probably would) but actually violates strict aliasing and is
thus technically "undefined behavior".

Having any undefined behavior at all is undesirable, and so a new
utility has been added that combats this by allowing for contiguous
sequences of floating point values to be interepreted as a sequence of
"T" types, such as "vector", "color", etc.

This also ensures better semantics are added to types, since we can now
be sure of the underlying types being used.