✨ Add vertex semantic type
In graphics development, it is commonly necessary to refer to vertex
positions while constructing models. Often, this is actually done by
reusing a "vector" type everywhere -- which generally confuses what
the underlying data is as a result.

It's much cleaner to represent this with a proper semantic type that is
meant to be usable in such a condition; and as a result, a "vertex" type
has been added that does this.

There is not a large amount of logic that is, or can be, associated with
it; it simply exists as a modeling primitive.