✨ Add texture_coordinate
All underlying graphics hardware has to perform some form of mapping
between images used for texturing, and the models that the textures
get applied to. This is commonly called a texcoord, or a texture
coordinate. Similar to vertices, this is often performed by reusing
utilities like a 2-component vector, rather than defining a full
type to represent the task properly.

This adds a type called texture_coordinate with the responsibility of
representing a texture coordinate, which is convertible to a 2-component
vector as-needed.