✨ Add 'normal' semantic type
3D graphics mathematics makes heavy use of `normal` vectors;
vectors which are normalized to unit length, that represent a normal
at a given `vertex`.

Since the semantics of a normal are similar to that of a vector, many
engines simply reuse the `vector3` type to implement this; however this
leads to the possibility of the invariant of the unit-length magnitude
to be violated due to incorrect mathematics. To prevent such a
possibility from occurring, a new semantic type named `normal` has been
added to the system which represents, and forces, normalized vectors
to be used. This type is easily convertible back to the basic
`vector3` type on-demand, allowing for simple interoperability.