🔨 Disable warnings for ignored attributes
The clang compiler is quite strict with its warning emissions for
attribute usage. Frequently, this is complaining due to functions being
defined with `inline` when being declared without it, due to the
`dllexport` attribute being used on the class.

Despite this warning, it has no practical negative effect. The generated
assembly and link maps are all correct, making this warning a pedantic
nuissance at best. This has been disabled to prevent this warning from
occurring.