🔨 Define larger matrix functions in source files
A large part of the matrix API was previously defined `inline` due to
the origin of the code beginning a template of any type `T`, which
naturally must also be `inline`.

However with recent modifications, this is no longer the case, and nor
is it truly necessary. A large amount of the matrix API generates large
code segments that can easily bloat an executable size. As a result,
any of the larger functions that are not worth inlining have been
relocated to the cpp file.