✨ Add unaligned utilities
When working with raw data coming from files, buffers, etc -- it can
frequenty be necessary to read data from an unaligned address as a
fundamental type like an integral value, floating point value, etc.

To allow for this to be possible in alloy, some new utilities have been
added that narrowly wrap `span` and `std::memcpy` so that values can be
ready from unaligned addresses in a simple way.

This is a proper testable practice that will also allow for higher-order
structures to be built off of this, such as a `packed_buffer` type which
will be necessary for storing mesh data for vertex attributes.