✨ Add packed_buffer data type
In a game engine, a large amount of data that is created, modified, or
just read will be in a packed binary format. This format does not align
well with the native alignment requirements of the types in which this
data represents.

One way of managing this is to simply `std::memcpy` the code each time,
but this isn't sustainable for large pieces of data that vary on the
underlying data-type, since the offset will be needed for each change.

To address these issues, a new packed_buffer data type has been added
that breaks the responsibilities up into several types:

* The `packed_buffer` represents a simple buffer bytes in a stream
* The `packed_buffer_writer` is a mutable view which allows for writing
  objects of different type directly to the end of the `packed_buffer`
* The `packed_buffer_reader` is an immutable view which allows for
  reading objects from the packed_buffer, optionally with support for
  seeking for the read location
* A concept for `Packer` types, which allow for user-defined custom
  means of packing and unpacking data from a tight stream. This allows
  a simple, configurable, and swappable means of configuring the same
  data-type from a stream. By default, everything is `std::memcpy`ed