🔨 Remove heterogeneous range from span
The span implementation, taken from BackportC++, has a bug in the
template implementation of the range constructor wherein the `End`
template argument is greedy. This is due to the under-constrained
nature of the constructor, which only participates in overload
resolution if the Iterator type is of the same value type, but not if
the end type is the correct value type.

Rather than checking both, this span implementation has just removed
heterogeneous iterator support entirely, since it's ultimately not
yet needed in this system. By the time heterogeneous iterators are
used, Alloy should be using C++20 and thus can use `std::span` directly.