✨ Add source_group for IDE projects
In CMake, certain generation targets must make use of the `source_group`
property to identify a logical source layout that is independent of
the physical source layout, otherwise the directory hierarchy is
ignored. This, in particular, applies to both Xcode and Visual Studio,
where the IDEs display all headers/sources in the same "directory"
unless `source_group` is applied.

To make development less painful on Xcode and Visual Studio, the various
CMakeLists have been updated to make use of `source_group` so that this
displays correctly.