✨ Organize cmake targets into folders
Some CMake generators offer the ability to graphically sort targets
based on the `FOLDER` attribute. This allows for the given IDE to
display all targets sorted under logical "folder"s, rather than
flattening all targets into a long list. This helps project organization
significantly since targets are easily discoverable.

The various CMakeLists have all been updated to now make use of this
property. This now sorts the following types of targets into the new
respective "folders":

* Main library (core, io, etc) -> "Component Targets"
* Extra libraries -> "Optional Component Targets"
* Example executables -> "Example Targets"
* Unit test executables -> "Unit Test Targets"
* Self-containment Tests -> "Containment Test Targets"

This organization is not concrete... It's likely that at some point
the various tests might be grouped together into:

* `Test Targets/`
  * `Unit Test Targets/`
  * `Containment Test Targets/`
* `Component Targets/`
  * `Required Component Targets/`
  * `Optional Component Targets/`

However, currently there is not a strong motivation for this due to the
lack of overall targets. Perhaps this might be added at some point in
the future.