🔨 Convert math constants to inline
When Alloy was first created, it was targeting C++14. As part of the
original design decisions, the `math_constants` class used to make
use of `constexpr` functions so that it could be defined easily in a
`constexpr` way without requiring an explicit storage backing.

However, shortly thereafter the project adopted c++17 -- which allows
for `inline` constants, which prevent the need for this.

The constants have been changed to make use of this syntax, and all
uses have been updated.