✨ Add operator[] to matrix classes
The matrix[2,3,4] classes were previously not designed to include
the `operator[]` function due to complications that prevent it from
behaving like a true 2d-matrix.

However, it appears that the simple answer is that this _is_ possible
so long as operator[] is implemented such that it returns a pointer to
the start of the row being acted on. Regardless of whether the data is
contiguously stored in a 1d or 2d array, this functions correctly -- and
generates nearly identical assembly in both cases.

As a result, this change has been added to provide a more idiomatic way
of accessing matrix functions.