🔨 Document `transform` with `[[nodiscard]]`
The C++17 attribute `[[nodiscard]]` was added to allow for the compiler
to generate useful warnings when a user calls a function where the
returned argument should not be ignored.

This is an extremely useful feature to prevent developer errors that
has been neglected from use within most of Alloy, but is something that
the project should be adopting more regularly.

The `transform` class has been updated to now contain these
documentations on the few places where it's applicable, as a starting
point that will continue to be done moving forward as well.