✨ Add lazy_transform
Transformation matrices may change frequently, but they are not
guaranteed to change every frame. For this reason, it's frequently
only ever necessary to compute the transformation matrices when
and only when these values have changed to save unnecessary compute
time.

To allow for this in a simple way, a new wrapper has been added:
lazy_transform. The previous transform types that have been added
all define the same `extract_matrix`, which is being leveraged by
this type to ensure that the extraction will only occur if the
underlying trasnformation has changed at any point.