✨ Add 'mesh' and 'mesh_builder'
All 3D models are comprised of a series of vertices, which describe
information about the model at each point:
* The location of the vertex,
* The normal at that point,
* The texture coordinate mapping of that vertex, and
* The base color of that model

This data can be quite large for large models, however it does not need
to be. Some parts of this data can easily be compressed, and vertices
can be reused:

* normals can be compressed to 16-bit integers, since the total number
  of valid inputs of a normalize [-1, 1] value is rather small
* color values can be compressed to 4-byte colors (or smaller)
* vertices can be reused multiple times using indices, to prevent the
  vertex from being duplicately defined

Additionally, this data can be done interlaced into a strided series
of data in the same stream. This last part allows for better
cache-locality per vertex which should enable faster throughput.

This commit introduces a `mesh` and `mesh_builder` type which make use
of the `packed_buffer` to interlace vertex data in this way.

It is possible that at some point in the future, this
`mesh`/`mesh_builder` may need to be refactored to support the option
for `mesh` being a container of `submesh`es as well. If this occurs,
the builder will likely be broken into a system that builds `submesh`
objects, and each submesh automatically attaches to the `mesh`.