🩹 Fix SFINAE constraints on quantity
The quantity class was merged a little earlier than it should have been
and unfortunately contained a few bugs that would only be visible
during use of the type. Most notable is that `quantity<T>` cannot
add to or subtract from other `quantity<T>` types, but rather can only
operate with `quantity<U>` types where `U` is related to `T`. This is
fundamentally a mistake, caused by an overly constrained SFINAE
restriction of `is_base_of<U,T>`.

These conditions have been changed to now check whether `T` is a base
of `U`, whether `U` is a base of `T`, or if `T` *is* `U`.

While reworking this, the formatting of `quantity` was updated to make
use of uniform function definitions as well, which has massively cleaned
up the nested `typename` identifiers.