🩹 Fix filesystem visibility and symbol exports
The `disk_filesystem` class mistakenly marked the `finalzie`
functionality as being `public`, when it is intended to be a `private`
and hidden part of the API.

The `file` class was mistakenly not documented with the `ALLOY_IO_API`
exporting symbol which prevented it from being used in outside projects.

These issues have been fixed.