✨ Add string tokenizer
The Alloy project recently came across a need to tokenize
space-delimited inputs in C++, but lacked any clear facilities to do
this. The closest thing to conventional for tokenization is to use
either iostreams, which are not (explicitly) used by this project, or
to use `std::strtok` which is horribly deprecated and should
not be used due to mutation of the input string.

This sparked the nead for a new `string_tokenizer` type that simplifies
this need. This type acts like a tokenizer, capable of breaking up
strings based on single-character delimiters