✨ Add string_utilities
Recent changes required a standardized way of parsing strings into a
numeric and floating point representations.
The standard C library provides some utilities that do this for us, but
the performance is typically sub-optimal, and requires explicitly
checking `errno` in order to determine error conditions, which is
undesirable. C++ offers some ways of doing this, but most require some
level of boilerplate in order to easily use -- such as `std::from_chars`,
which requires checking each return type for success.

This introduces parsing boilerplate that wraps `from_chars` where
possible for integer conversions, and `std::stroto[f|d]` for floats,
due to an omission from the standard library for floating point types.