✨ Add monadic utilities to `expected`
With recent change to parsing having been added to the project which
return `expected` objects, it's extremely helpful to have higher-level
utilities that act on these objects in a coherent and functional way.

For that reason, new functional wrappers have been added that allow for
the `expected` API to be propagated with simple `map`/`flat_map`/
`and_then`/`value_or` functionality. This can allow badly parsed data
to provide reasonable defaults on failure, and to allow the data to be
streamed and propagate the error along.