✨ Add ostream utility for vectors
There are various times during debugging that it would be convenient to
be able to simply print 2, 3, and 4 component vector objects.
Alloy does not make use of iostreams anywhere in the current
implementation since simply including the `<iostreams>` header
introduces static overhead, which is undesirable. However, iostreams
are useful during debugging purposes.

As a compromise, to allow for iostream usage of vectors while also
avoiding the overhead, a separate header has been added that provides
the ostream definitions as inline definitions. This ensures that the
user must explicitly request it in order to pay for it.