🔨 Add `comparison_tolerance` to vector classes
The various vector classes have been given a new static member,
`comparison_tolerance`, to use as the default tolerance for the
vector classes.

By default this value will be the same as the `default_tolerance`, but
in the future may be redefined to be more something more or less
restrictive, based on the needs of the system.