🔧 Add Catch as submodule
Working with `conan` to handle all dependencies has become a hassle
for the CI workflow. `conan` requires knowing a lot of information about
the current architecture, even if the dependency is header-only -- which
is completely unnecessary.

As useful as `conan` is, it overcomplicates the system and adds an
external dependency to the codebase. Since this is (currently) only
used for the `catch` dependency, its easier to simply add this as either
a subtree or submodule instead. The latter approach has been chosen here
so that it may simply be cloned with a basic `git clone --recursive`
command.

Subtree could also have been a viable alternative, however for the
purposes being used here would likely have been overkill, and would be
checking in a rather large chunk of code for the sole purpose of being
used as a dependency.