🩹 Fix default `expected` constructors
When `= default`ing constructors, it can be detrimental to specify
the `noexcept`-ness and `constexpr`-ness of the constructor since it
may either be misleading, or worse, implicitly deleted if the
`noexcept`ness cannot be satisfied.

GCC has determined that `expected() noexcept = default` is resulting in
an implicitly deleted default-constructor due to the constructor not
actually being `noexcept`. The exact reason for this failure is not
known, but realistically, the non-throwing nature of the constructor
should *not* be cause for this type of issue.

As a result, all `expected` constructors have removed `constexpr` or
implicitly expected `noexcept`-ness to instead simply be determined
by the compiler with `= default`.