🧹 Move zstring_view to its own separate header
The `zstring_view` type was previously coupled with the `string_view`
type's definition. Although this is, in general, a useful utility --
it is more convenient to have these definitions split independently
so that the types' origins are more easily identified.

As a result, the `zstring_view` has been moved to its own separate
header.