🩹 Suppress `-Wswitch-enum` on `clang-cl`
`-Wswitch-enum` warns on any unhandled enumeration, including in
any enum that falls through with a `default` statement.
It's completely unreasonable to handle all cases of an enum in cases
where `default` is for the intended purpose -- and especially when the
underlying enumeration is from the standard library, such as `std::errc`
(the case triggering this error).

This warning, and `-Werror`, has been disabled