✨ Update delegate to support better binding
The `delegate` type was previously only bindable by using `make` or
`bind` calls, which was rather limiting in that it requires the full
name of the `delegate` type in order to produce a proper delegate
object.

This modifies the signature to be based off of the bitwizeshift/delegate
library instead -- which allows for `core::bind<...>(...)` expressions
to return an object that is convertible to a `delegate`.