✨ Add `add_vertices` to `mesh_builder`
The `mesh_builder` class previously was only capable of adding singular
vertices at a time -- which requires several operations at a time:

* acquire current size (for the vertex index),
* pack the vertex, and
* Increment the vertex count

This leads to unnecessary cost/cycles when done in a loop from a series
of known vertices.
To address this, new `add_vertices` and `add_faces` functions were
added that don't perform any incrementing to ensure that the packing
operations are more optimal.