🩹 Remove extern templates in attribute_buffer_view
The `attribute_buffer_view` file previously was making use of
`extern template` syntax to minimize the cost of compile-time template
instantiations in consumer code. However, this appears to conflict
greatly with the varying implementations of `dllimport`/`dllexport` on
Windows.

On MSVC, this works fine as long as the symbols are marked as
`dllexport`. On `gcc`, this generates an error but *sometimes* works.
On `clang`, this consistently fails with unresolved symbol errors.
Note that this is not referring to `-fvisibility=...` which is similar
for GNU-based systems; this is purely with Windows DLL formats.

Due to this common failure, quite often any code on Windows has required
custom modifications to this file in order to continue working on the
Alloy project -- which really has been both frustrating any undesirable.

For the time being, the `extern template` has been removed -- and may
be revisited in the future if these templates actually pose a real
compile-time problem in practice.