✨ Add game loop utilities
Games require some form of game-loop that continuously executes all
physics integration and rendering. This can be complicated to get
correct, especially when it comes to creating a variable vs fixed
time-step game-loop.

This implements both a semi-fixed and completely-fixed game loop that
is implemented using the CRTP to avoid potential virtual calls.