✨ Add specialization support to `interpolation`
The `interpolation` utilities were previously implemented purely in
terms of `operator+` and `operator*` -- using the standard lerp
equation.

The downside to this is that not all objects will implement these
operations since they don't all logically lend themselves to the API.
For example, the `transform` object does not make sense to define
support for `*` or `+`, but it does make sense to be able to lerp
between the two. This now allows for specializations to perform this
instead.