✨ Add `interpolated_transform`
When using a fixed-time-step game loop, it's important to be interpolate
the state between the last physics-adjustment and the current expected
state depending on where the frame time is relative to when the render
time will occur. This is to ensure that we don't have any unsightly
and unexpected/jarring graphical effects, which can occur if the
frame-rendering-time is slightly out-of-sync with the physics adjustment
step.

This implements a wrapper around transformations, similar to the
`lazy_transform`, that computes the interpolated state between the
previous and current "logical" state in order to determine the exact
"frame" state for rendering.