📌 Add dependency to TGA library
Game engines fundamentally require _some_ way of loading image
resources. A common file type used in game engines is TGA, since the
majority of the data is simple and quick to read, easy to write, and
uncompressed (thus requiring few third-party dependencies).

Since Alloy will, at some point, take on the responsibility of image
reading, a dependency has been added to the TGA library in order to
support this. This may be removed at a later point in time, once TGA
is supported in Alloy directly.