🔨 Return `eof` from `file::read`
The `file::read` API previously never returned `eof` as a status, which
can result in potential bugs when coupled with the `file_serializer`
API which makes heavy use of the reading.

This fixes it so that `eof` is regarded as an "error" only when no bytes
are read from the file (but not necessarily when only a partial amount
are read).