✏️ Rename `point` to `point3`
All geometry in Alloy to date has been 3-dimensional, completely
dropping 2D logic anywhere. For the most part this is fine; however
this becomes problematic when considering graphics engines that need to
support both 3D and 2D (for example, to represent billboard or text
locations).

Due to the 3D geometry, the `point` class previously represented only a
3D point, and no such 2D point was present; however it's becoming
increasingly apparent that logic for 2D is still also necessary in
addition to 3D. For this reason, the `point` class has been renamed to
be `point3` instead.