✨ Add `polar` point geometric type
Sometimes when dealing with level geometry, it may be useful to operate
in terms of polar a coordinate system rather than straight cartesian.
A simple example of where this can prove beneficial is when an object
is meant to circle another object. Polar coordinates can simplify this
process by ensuring the coordinate system is consistent.