✨ Add `for_each_member` utility
It is frequently beneficial and desirable to be able to iterate all
members of a given struct; though the C++ language offers no real
facility to do this. Often, this results in users needing to manually
author code to perform this simple task, which can be both undesirable
and overall inefficient in terms of developer work-hours.

To compensate for this, a library solution has been added to Alloy that
leverages SFINAE and structured-binding to access the underlying types
stored within a `struct` object.