✏️ Rename `copy(...)` to `copy_with(...)`
Any container that uses heap memory within Alloy defines an explicit
`copy()` function to perform copies, so that users don't unintentionally
waste cycles or heap memory performing copies manually.

The benefit of this is that we establish better type-safety to prevent
unintended copies statically while also preserving the same performance
as a regular old copy-constructor (thanks to C++17's guaranteed
copy-elision).

For copies that also establish a new allocator, this was previously
implemented by having a `copy(...)` function that took the underlying
allocator. However, this overload is not quite explicit what its intent
is; and so this has been renamed to be `copy_with(...)` instead.