✨ Move and re-implement `mesh` in `render`
The `mesh` type was previously implemented under Alloy::Core. This type
was intended to be "generic" in the conventional sense -- but overall
hits issues once you attempt to implement various things like
"materials" (which are coupled to meshes), or certain render features
like topology or winding order.

Ultimately, it seems to be a better fit to move all physical geometry
to Alloy::Render rather, and leave the logical geometry under
Alloy::Core. This divide overall helps the general SoC of the project,
since many of these features will require intricate ties with the
underlying renderer.

In the process of this, the various attribute types have been removed.