✨ Add generic 2D and 3D polygon mesh builders
It can sometimes be useful to offer a generic builder for polygons --
whether this is as a developer utility, or as a means to dynamically
produce objects at runtime. Although there exists the
`unsafe_mesh_builder` type previously introduced -- this type does not
uphold any of its required invariants, and is thus easy to break when
used by the end-user.

This adds a new polygon mesh builder that simplifies this process by
intentionally forcing some of these invariants to be upheld. The
2D one assumes 2-dimensional points, whereas the 3-dimensional one
assumes 3D points that may contain optional normals, tangents, and
bitangents.