✨ Add gl state-cache
It can be expensive to call OpenGL functions, especially when changing
shader-related state such as depth-processing, face-culling, etc.
This becomes a wasted and unnecessary cost especially when invoking
these functions with the same state we are already in.

To prevent potential issues, a state cache has been added that
records the most recent known OpenGL state, and only invokes the
change functions when necessary. This will be used in a
higher-composition later in the renderer.