✨ Add `color` support for meshes
The `mesh` object previously was written solely in terms of
positions, uvs, and tangent-space arguments -- but no support for color.

It turns out that color can be useful for debugging, for example when
trying to color basic geometry without textures (such as an axes or
a grid). This adds support for this new feature.