✨ Move `light` types from `core` to `render`
The `light` types were previously defined under `alloy/core/model`, but
with the recent transition of mesh types to `render`, the lights are
better suited to this project.

This moves the light types to Render. Additionally, any equality
functions have been removed, as this functionality is unlikely to really
be beneficial.