✨ Add `core::optional`
C++17's `std::optional` type is extremely useful for representing
nullability, but it only extends as far as values -- but not
references. Although a "nullable reference" in C++ is otherwise known
as a pointer, having an actual `optional<T&>` has several advantages:

* A consistent API for nullability
* Chainable monadic functionality, otherwise not possible with raw
  pointers
* Interoperability with `optional<T>` value types, such as constructing
  an `optional<T>` from an `optional<const T&>`, or similarly
  performing comparisons

Ultimately, this is beneficial to have -- and so this has been added
by simply inheriting from `std::optional` and augmenting/adding some
functionalities.

Additionally, utilities for converting between `optional` and `result`
have been added as well.