🔨 Support local and world camera translations
The `camera` class has learned how to perform local translations
relative to the direction the camera is facing, to allow for forward
movements.

The old translation functionalities have been renamed to be a
`translate_world*` series of functions, whereas the new ones are labeled
`translate_local*`.

This feature now allows for cameras to be moved relative to the
direction it faces.