✨ Add `enum_traits` for iteration support
Enums in C++ are very lacking in overall functionality. Its not possible
to iterate over all enums, and its also not possible to know the first
and last value in said enums.

In an effort to simplify this, the logic has been extracted into traits
to determine contiguity of values, along with the max enum value.