✨ Add `option_set` utility
The previous change introduced a new utility for statically indicating
the reachability of `enum` types. With this functionality, its now
feasible to implement a `bitset`-like way to test `enum` values for
being set, while retaining the type-safety of the enum.

This effectively replaces the need to manually calculate (or define)
flag values as fixed bits -- which allows more flexibility and
ensures correctness. This has been added as a typed called `option_set`,
named after the Swift iOS equivalent of the same name, and will now
function as the equivalent of bit masks for enumerators.