🧹 Suppress bad `-Wdtor-name` diagnostic
Modern versions of clang are complaining about
`result_union<T,E,false>::~result_union`, claiming that this is not the
same name as the dtor and is thus a violation of the ISO C++ spec.

However, this _is_ the correct name of the destructor since this is a
partial specialization and not a full specialization (thus
`~result_union<T,E,false>()` is not valid).

This warning has been suppressed.