✨ Add `types` and `int_utilities`
Most of Alloy currently operates on the `cstdint` header for the types;
however it would be nice to have shorter aliases that can be used
throughout the systems to save on typing. These types have been
realiased as `uN` and `sN` for `uintN_t` and `intN_t` respectively.

Additionally, utilities have been added to aid in the construction and
decomposition of these integral types. These are effectively integer
factories to ensure easier construction.

Tests have been added to ensure the behavior is correct, as expected.