🧹 Clean up `color` with new `int_utilities`
A lot of the `color` logic for `rgba` construction from 8 or 16 bit
values has recently been duplicated in the `int_utilities` support.
This duplication promotes better code-sharing and minimizes the number
of lines required to achieve the same output.

The color class has been updated to now benefit from these changes, which
also now updates many functions to be `inline` and `constexpr` rather
than defined in the `cpp` file.