✨ Generate default `almost_equal`
The current implementation of `almost_equal` requires manually defining
both an `almost_equal(lhs, rhs, tolerance)` and an
`almost_equal(lhs,rhs)` -- the latter of which is almost always an exact
duplicate of the former, simply passing a tolerance in instead.

Ultimately, this process is mostly unnecessary and can be automated.
This has been changed to now automatically generate the `lhs,rhs` form
without a specific `tolerance` that will automatically use either the
`default_tolerance` if specified, or `T::comparison_tolerance`.

This will call the 3-component `almost_equal` as discovered by ADL.