✨ Add `polymorphic_vector`
Standard containers in the C++ library do not natively support
polymorphism. Ordinarily, this is solved by explicitly storing
either `unique_ptr` or `shared_ptr` objects -- but this doesn't
work well for cases where the allocator being used in the `vector`
and the allocator being used in the pointers are the same -- since
the deleter will require this information.

To combat this, a `polymorphic_vector` was added. This is effectively
a `vector` that supports polymorphic data types in a coherent way,
all that share the same `allocator`